Add tests for deduceForkPoint, extract new common tests methods to TestUtils
