+++ backendRoot/src/test/java/com/virtuslab/gitmachete/backend/root/GitMacheteRepositoryBuilder_deriveSyncToParentStatusTest.java
@@ -1,10 +1,10 @@
 package com.virtuslab.gitmachete.backend.root;
 
-import java.nio.file.Path;
-import java.util.Date;
-import java.util.Optional;
+import static com.virtuslab.gitmachete.backend.root.TestUtils.TestGitCoreRepository;
+import static com.virtuslab.gitmachete.backend.root.TestUtils.TestGitCoreRepositoryFactory;
+import static com.virtuslab.gitmachete.backend.root.TestUtils.getCommit;
+import static com.virtuslab.gitmachete.backend.root.TestUtils.getGitCoreLocalBranch;
 
-import io.vavr.collection.List;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -12,13 +12,8 @@ import org.powermock.api.mockito.PowerMockito;
 import org.powermock.reflect.Whitebox;
 
 import com.virtuslab.gitcore.api.BaseGitCoreCommit;
-import com.virtuslab.gitcore.api.BaseGitCoreCommitHash;
-import com.virtuslab.gitcore.api.GitCoreException;
 import com.virtuslab.gitcore.api.IGitCoreBranch;
 import com.virtuslab.gitcore.api.IGitCoreLocalBranch;
-import com.virtuslab.gitcore.api.IGitCorePersonIdentity;
-import com.virtuslab.gitcore.api.IGitCoreRemoteBranch;
-import com.virtuslab.gitcore.api.IGitCoreRepository;
 import com.virtuslab.gitmachete.backend.api.SyncToParentStatus;
 
 public class GitMacheteRepositoryBuilder_deriveSyncToParentStatusTest {
@@ -34,6 +29,17 @@ public class GitMacheteRepositoryBuilder_deriveSyncToParentStatusTest {
   @BeforeClass
   public static void init() {
     Whitebox.setInternalState(gitMacheteRepositoryBuilder, "gitCoreRepositoryFactory", repositoryFactory);
+  }
+
+  SyncToParentStatus invokeDeriveSyncToParentStatus(IGitCoreLocalBranch coreLocalBranch,
+      IGitCoreBranch parentCoreLocalBranch,
+      BaseGitCoreCommit forkPoint) throws Exception {
+    return Whitebox.invokeMethod(gitMacheteRepositoryBuilder,
+        "deriveSyncToParentStatus",
+        repository,
+        coreLocalBranch,
+        parentCoreLocalBranch,
+        forkPoint);
   }
 
   @Test
@@ -70,7 +76,8 @@ public class GitMacheteRepositoryBuilder_deriveSyncToParentStatusTest {
     BaseGitCoreCommit parentPointedCommit = getCommit(null);
     BaseGitCoreCommit branchPointedCommit = getCommit(parentPointedCommit);
     IGitCoreLocalBranch parent = getGitCoreLocalBranch(parentPointedCommit);
-    IGitCoreLocalBranch branch = getGitCoreLocalBranch(branchPointedCommit, /* forkPoint */ parentPointedCommit);
+    IGitCoreLocalBranch branch = getGitCoreLocalBranch(branchPointedCommit,
+        /* forkPoint */ parentPointedCommit);
 
     // when
     SyncToParentStatus syncToParentStatus = invokeDeriveSyncToParentStatus(branch, parent, parentPointedCommit);
@@ -125,153 +132,5 @@ public class GitMacheteRepositoryBuilder_deriveSyncToParentStatusTest {
 
     // then
     Assert.assertEquals(SyncToParentStatus.OutOfSync, syncToParentStatus);
-  }
-
-  static TestGitCoreCommit getCommit(BaseGitCoreCommit parentCommit) {
-    assert parentCommit == null || parentCommit instanceof TestGitCoreCommit;
-    return new TestGitCoreCommit((TestGitCoreCommit) parentCommit);
-  }
-
-  SyncToParentStatus invokeDeriveSyncToParentStatus(IGitCoreLocalBranch coreLocalBranch,
-      IGitCoreBranch parentCoreLocalBranch,
-      BaseGitCoreCommit forkPoint) throws Exception {
-    return Whitebox.invokeMethod(gitMacheteRepositoryBuilder,
-        "deriveSyncToParentStatus",
-        repository,
-        coreLocalBranch,
-        parentCoreLocalBranch,
-        forkPoint);
-  }
-
-  IGitCoreLocalBranch getGitCoreLocalBranch(BaseGitCoreCommit pointedCommit)
-      throws GitCoreException {
-    return getGitCoreLocalBranch(pointedCommit, null, false);
-  }
-
-  IGitCoreLocalBranch getGitCoreLocalBranch(BaseGitCoreCommit pointedCommit, BaseGitCoreCommit forkPoint)
-      throws GitCoreException {
-    return getGitCoreLocalBranch(pointedCommit, forkPoint, false);
-  }
-
-  IGitCoreLocalBranch getGitCoreLocalBranch(BaseGitCoreCommit pointedCommit, BaseGitCoreCommit forkPoint,
-      boolean hasJustBeenCreated)
-      throws GitCoreException {
-    IGitCoreLocalBranch mock = PowerMockito.mock(IGitCoreLocalBranch.class);
-    PowerMockito.doReturn(pointedCommit).when(mock).getPointedCommit();
-    PowerMockito.doReturn(Optional.ofNullable(forkPoint)).when(mock).deriveForkPoint();
-    PowerMockito.doReturn(hasJustBeenCreated).when(mock).hasJustBeenCreated();
-    return mock;
-  }
-}
-
-class TestGitCoreCommit extends BaseGitCoreCommit {
-  private static int counter;
-
-  private final int id;
-
-  private final TestGitCoreCommit parentCommit;
-
-  TestGitCoreCommit(TestGitCoreCommit parentCommit) {
-    this.parentCommit = parentCommit;
-    id = counter++;
-  }
-
-  TestGitCoreCommit getParentCommit() {
-    return parentCommit;
-  }
-
-  @Override
-  public String getMessage() {
-    return null;
-  }
-
-  @Override
-  public IGitCorePersonIdentity getAuthor() {
-    return null;
-  }
-
-  @Override
-  public IGitCorePersonIdentity getCommitter() {
-    return null;
-  }
-
-  @Override
-  public Date getCommitTime() {
-    return null;
-  }
-
-  @Override
-  public BaseGitCoreCommitHash getHash() {
-    return new BaseGitCoreCommitHash() {
-      @Override
-      public String getHashString() {
-        return String.valueOf(id);
-      }
-    };
-  }
-}
-
-class TestGitCoreRepository implements IGitCoreRepository {
-  @Override
-  public boolean isAncestor(BaseGitCoreCommit presumedAncestor, BaseGitCoreCommit presumedDescendant) {
-    assert presumedAncestor instanceof TestGitCoreCommit;
-    assert presumedDescendant instanceof TestGitCoreCommit;
-    while (presumedDescendant != null && !presumedDescendant.equals(presumedAncestor)) {
-      presumedDescendant = ((TestGitCoreCommit) presumedDescendant).getParentCommit();
-    }
-    return presumedDescendant != null;
-  }
-
-  @Override
-  public Optional<IGitCoreLocalBranch> getCurrentBranch() {
-    return Optional.empty();
-  }
-
-  @Override
-  public IGitCoreLocalBranch getLocalBranch(String branchName) {
-    return null;
-  }
-
-  @Override
-  public IGitCoreRemoteBranch getRemoteBranch(String branchName) {
-    return null;
-  }
-
-  @Override
-  public List<IGitCoreLocalBranch> getLocalBranches() {
-    return List.empty();
-  }
-
-  @Override
-  public List<IGitCoreRemoteBranch> getRemoteBranches() {
-    return List.empty();
-  }
-
-  @Override
-  public Path getRepositoryPath() {
-    return null;
-  }
-
-  @Override
-  public Path getGitDirectoryPath() {
-    return null;
-  }
-}
-
-class TestGitCoreRepositoryFactory implements IGitCoreRepositoryFactory {
-
-  private final TestGitCoreRepository instance;
-
-  public TestGitCoreRepositoryFactory() {
-    instance = new TestGitCoreRepository();
-  }
-
-  public TestGitCoreRepository getInstance() {
-    return instance;
-  }
-
-  @Override
-  public IGitCoreRepository create(Path pathToRoot) {
-    return getInstance();
   }
 }
